requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0

# Image Processing
scipy>=1.11.0
matplotlib>=3.7.0
tqdm>=4.65.0

# Additional
streamlit>=1.25.0
